// src/model/productEntity.ts
export interface ProductEntity {
    AanbodRegels: {};
    ArtikelGroep: {};
    ArtikelGroepNaam: string;
    Beladingen: {};
    Code: string;
    Eigenschappen: {};
    Fotos: {};
    Hoogte: string;
    ID: number;
    Naam: string;
    Potmaat: string;
    PotmaatNumeriek: number;
    VbnProductCode: string;
    VbnProductNaam: string;
}